requirements.txt
absl-py==2.1.0
altgraph==0.17.4
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.1.1
cv2_enumerate_cameras==1.1.17
cycler==0.12.1
flatbuffers==24.3.25
fonttools==4.55.0
getmac==0.9.5
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
jax==0.4.13
kiwisolver==1.4.7
matplotlib==3.7.5
mediapipe==0.10.11
ml-dtypes==0.2.0
numpy==1.24.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
protobuf==3.20.3
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
requests==2.32.3
scipy==1.10.1
six==1.16.0
sounddevice==0.5.1
urllib3==2.2.3
zipp==3.20.2
